Extract the subscribe loop into a separate mod
